rest_django_trading/requirements.txt
Django>=2.0,<3.0
gunicorn==19.9.0
psycopg2>=2.7,<3.0
djangorestframework==3.11.0
psycopg2-binary==2.8.2
requests==2.23.0